+++ src/Page/MyPage/MyPage/scss/chargeBanana.module.scss
@@ -1,0 +1,209 @@
+.modalContainer {
+  width: 100%;
+  height: 100%;
+  flex-shrink: 0;
+  background: rgba(0, 0, 0, 0.5);
+  position: fixed;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  right: 0;
+  top: 0;
+
+  .modalCard {
+    display: flex;
+    width: 320px;
+    height: 512px;
+
+    flex-direction: column;
+
+    flex-shrink: 0;
+    border-radius: 14px;
+
+    background: var(--GRAY-SCALE-Gray-900_CARD, #2b2c31);
+
+    /* card shadow1 */
+    box-shadow: 0px 0px 16px 0px rgba(0, 0, 0, 0.1);
+
+    .modalHeader {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+
+      .modalHeaderText {
+        color: var(--GRAY-SCALE-Gray-200_DEFAULT-MAIN-TEXT, #eff0f4);
+
+        /* SUBTITLE/16_bold */
+        font-family: Pretendard;
+        font-size: 16px;
+        font-style: normal;
+        font-weight: 700;
+        line-height: 160%; /* 25.6px */
+        padding: 11px;
+        width: 230px;
+        text-align: center;
+        margin-left: 30px;
+      }
+      .close {
+        cursor: pointer;
+        margin-bottom: 3px;
+      }
+    }
+
+    .hr {
+      width: 100%;
+      border-top: 1px solid var(--font-color-gray-700);
+    }
+
+    .contentContainer {
+      padding: 0.5rem;
+
+      .modalImg {
+        display: flex;
+
+        flex-shrink: 0;
+        margin-top: 1rem;
+        margin-bottom: 1rem;
+
+        img {
+          width: 80px;
+          height: 80px;
+
+          margin: 0 auto;
+        }
+      }
+
+      .modalText {
+        margin-bottom: 2rem;
+        color: var(--GRAY-SCALE-Gray-200_DEFAULT-MAIN-TEXT, #eff0f4);
+        text-align: center;
+
+        /* BODY/12_medium */
+        font-family: Pretendard;
+        font-size: 12px;
+        font-style: normal;
+        font-weight: 500;
+        line-height: 160%; /* 19.2px */
+      }
+
+      .bananaContainer {
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+
+        width: 100%;
+        height: 46px;
+        padding: 0 0.7rem;
+
+        // width: 301px;
+        // height: 46px;
+        // padding: 10px;
+        // gap: 10px;
+        flex-shrink: 0;
+
+        border-radius: 5px;
+        cursor: pointer;
+
+        &:hover {
+          background: var(--GRAY-SCALE-Gray-700_DISABLE, #37383e);
+          .numOfBanana {
+            .button {
+              border: 1px solid #2458ec;
+            }
+          }
+        }
+
+        .numOfBanana {
+          display: flex;
+          align-items: center;
+
+          gap: 8px;
+
+          .button {
+            display: flex;
+            align-items: center;
+            justify-content: center;
+
+            width: 20px;
+            height: 20px;
+            border-radius: 50%;
+            border: 1px solid #75777f;
+            box-sizing: border-box;
+
+            &.clicked {
+              border: 1px solid #2458ec;
+
+              div {
+                width: 14px;
+                height: 14px;
+
+                border-radius: 50%;
+
+                background: #2458ec;
+              }
+            }
+          }
+          .banana {
+            color: var(--GRAY-SCALE-Gray-200_DEFAULT-MAIN-TEXT, #eff0f4);
+
+            /* BODY/12_medium */
+            font-family: Pretendard;
+            font-size: 12px;
+            font-style: normal;
+            font-weight: 500;
+          }
+          .bananaNum {
+            color: var(--Lemon-Lemon-400, #d2f75a);
+
+            /* SUBTITLE/14_bold */
+            font-family: Pretendard;
+            font-size: 14px;
+            font-style: normal;
+            font-weight: 700;
+          }
+        }
+        .price {
+          color: var(--GRAY-SCALE-Gray-100, #f9fafe);
+          text-align: right;
+
+          /* SUBTITLE/16_bold */
+          font-family: Pretendard;
+          font-size: 16px;
+          font-style: normal;
+          font-weight: 700;
+        }
+      }
+
+      .payment {
+        margin-top: 24px;
+        display: flex;
+        width: 280px;
+        height: 38px;
+        padding: 7px 16px;
+        justify-content: center;
+        align-items: center;
+        gap: 4px;
+        flex-shrink: 0;
+        border-radius: 5px;
+        background: var(--GRAY-SCALE-Gray-700_DISABLE, #37383e);
+        color: var(--GRAY-SCALE-Gray-500, #93959d);
+        text-align: center;
+
+        /* SUBTITLE/16_bold */
+        font-family: Pretendard;
+        font-size: 16px;
+        font-style: normal;
+        font-weight: 700;
+        line-height: 160%; /* 25.6px */
+
+        &.clicked {
+          background: var(--Blue-Blue-400, #2458ec);
+          color: #fff;
+          cursor: pointer;
+        }
+      }
+    }
+  }
+}
